+++ CHANGELOG.md
@@ -11,6 +11,8 @@ SPDX-License-Identifier: LGPL-2.1-or-later
 * Remove deprecated `Mark` and `Release`. Use `MarkTmpMem` and
   `ReleaseTmpMem` instead.
 
+* Remove deprecated `HeapAllocMode` and its flags `FROM_BOTTOM` and `FROM_TOP`.
+  They lost any influence, as we use the system heap.
 # dune-uggrid 2.9 (2022-11-25)
 
 * The `dune-uggrid` module does not set the preprocessor flag `HAVE_UG` anymore.